Pre-Alpha 3.4
Added mouseclick support for spells, and bashing attacks, and simplified
the mouseclick process for charging and shooting your bow at things. I
also made it easier to add mouseclick support to special attacks in the
future.

Fixed up a bunch of stuff in the game, I made the sight algorithm work
better (potentially? I haven't had a chance to test this particular
iteration yet because my computer is being slow and I want to go to bed)
and I tightened up the use_icobar functions - by which I mean that I got
rid of the 5 partial ones and now only the full one exists.

Oh and the other 3 characters are now completely in the game, wireframes
and all.