Pre-Alpha 4.1
Smoothed out Talisman stuff. Added a bunch of new features, like the
ability to win the game. Boneyard is fully implemented now and is
waiting for your on floor 12. You won't be able to leave until you kill
it. Good luck! Helmets are also in, but you can't repair them yet.