Pre-Alpha 4.2
Tore out a whole bunch of the shop code and redid it. Long story short
adding new items to the game should be much easier now and won't
actually increase the amount of memory used after a few more tweaks
involving a content manager. Also the shops are way, way, waaaay better
and no longer give the player a set of items that they can't afford.