Pre-Alpha 4.22
Some new prompts for the shop are in. Added some small refinements on
the dungeon generation algorithm, so the player won't be swarmed by
monsters as soon as they enter the dungeon any more. This is the last
feature from the C++ version that I put in. Added the ability to destroy
altars.