Pre-Alpha 4.32
Dungeons can parse theme instructions. Added a "naturalfeature"
generator that can make rivers currently and will be extended to make
lakes, chasms, etc. Expanded the room class a little bit to take
advantage of future changes to the editor and to make code more
readable. More on that front to come in the near future.